Preserve selection when transitioning to vimpulse-visual-mode
Previously, when using emacs-selection commands in vi-mode that leave
the cursor on the right, an extra character would be effectively
selected because of the way vimpulse includes the character to the right
of the point as selected.

This commit adjusts the point by moving it to the left by one character
in these cases, so commands such as shift-meta-f will select to the end
of the word, instead of the end + 1.